Added namespaces and manual class mappings
Now you can add a namespace to a distinct langClassName, which helps
packages in laravel to provide model key translations.

The second thing that was added is the possibility to manually map a
class to a langClassName. This is helpfull in situations where you
overwrite a class with a different name and dont want to copy the
translationn entries.